Cleaned stuff up a bit and added SnapToDegree.java

+++ src/main/java/frc/team4931/robot/commands/utilities/Wait.java
@@ -12,9 +12,9 @@ import edu.wpi.first.wpilibj.command.Command;
 public class Wait extends Command {
 
   private long startTime;
-  private short waitTime;
+  private long waitTime;
 
-  public Wait(short waitTime) {
+  public Wait(long waitTime) {
     this.waitTime = waitTime;
   }
